requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tqdm==4.65.0
pyyaml==6.0   # Changed from 6.0.1 to 6.0 to resolve dependency conflict
requests==2.31.0
python-dateutil==2.8.2

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Reinforcement Learning
stable-baselines3==2.1.0
gym==0.26.2

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2

# Visualization
plotly==5.16.1
dash==2.13.0

# Data sources
python-binance==1.0.19
pybit==2.4.1
kucoin-python==1.0.11
yfinance==0.2.28
alpaca-trade-api==3.0.2

# Utilities
pytz==2023.3
apscheduler==3.10.4
tabulate==0.9.0
joblib==1.3.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0